% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fattvpVAR-package.R
\docType{package}
\name{fattvpVAR-package}
\alias{fattvpVAR-package}
\alias{fattvpVAR}
\title{Bayesian VAR with Stochastic volatility, asymmetry and fat tails}
\description{
Bayesian VAR with Stochastic volatility, asymmetry and fat tails
}
\details{
\tabular{ll}{
Package: \tab fattvpVAR\cr
Type: \tab Package\cr
Version: \tab 0.2.0 \cr
Date: \tab 2021-03-29\cr
License: \tab GPL (>=2)\cr
Depends: \tab R (\eqn{\geq 3.5.0}{>= 3.5.0})\cr
Imports: \tab Rcpp \cr
Suggests: \tab  \cr
LazyLoad: \tab yes }
}
\references{

}
\author{
Hoang Nguyen
}
