% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{get_forecast}
\alias{get_forecast}
\title{Forecast future observations of tvpVAR model}
\usage{
get_forecast(Chain, y0 = NULL, t_pred = 24, Nfsample = NULL)
}
\arguments{
\item{Chain}{The TVP-VAR-SV object from command fitTVPGaussSV/fitTVPStudentSV}

\item{t_pred}{The time prediction horizon.}
}
\value{
The list of forecast future observations of BVAR model
}
\description{
This function returns a forecast future observations of TVP-VAR-SV model.
}
\examples{
\dontrun{
forecast1 <- get_forecast(Chain1)
}

}
