% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEnt_TVPGSV.R
\name{MLEnt_TVPGSV}
\alias{MLEnt_TVPGSV}
\title{Marginal likelihood using Entropy method for TVP VAR model with Gaussian SV}
\usage{
MLEnt_TVPGSV(Chain, numCores = 4)
}
\description{
Marginal likelihood using Entropy method for TVP VAR model with Gaussian SV
}
