% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simtvpVAR.R
\name{sim.tvpVAR.nonSV}
\alias{sim.tvpVAR.nonSV}
\title{Simulate data from TVP VAR model without SV}
\usage{
sim.tvpVAR.nonSV(
  dist,
  K = 3,
  p = 2,
  t_max = 1000,
  b0 = 0.5,
  a0 = 0.5,
  h = -3,
  sigma_ab = NULL,
  y0 = matrix(0, ncol = K, nrow = p),
  nu = 6,
  is_tv = rep(0, K),
  seednum = 0,
  burn_in = 0
)
}
\arguments{
\item{dist}{The variable specifies the VAR error distribution. It should be one of
c("Gaussian","Student").}

\item{K}{The number of variables in VAR model.}

\item{p}{The number of lags in VAR model.}

\item{t_max}{The number of observations}

\item{b0}{The coefficients of B matrix.}

\item{a0}{The coefficients of A matrix.}

\item{h}{The initial log diag(Sigma_t) matrix. sigma_h is set at diag(0.01, K)}

\item{nu}{The degree of freedom.}

\item{seednum}{The default seed.}

\item{burn_in}{The discarded observations.}
}
\value{
A list of simulated data with its specification.
}
\description{
This function simulate data from a TVP VAR model without SV.
\deqn{A_t y_t = B_t x_t + SQRT(w_t) H eps_t}
}
\examples{
\dontrun{
datagen <- sim.tvpVAR.SV(dist="Gaussian")
y <- datagen$y
}
}
